adding new currency from bottom sheet using retrofit

+++ app/src/main/java/com/example/homework_recyclerview/domain/repository/Currency.kt
@@ -5,8 +5,8 @@ import com.example.homework_recyclerview.domain.repository.Parent
 
 data class Currency(
     override val id: Int,
-    var text: Int,
+    var text: Double,
     val type: String,
     @DrawableRes val flag: Int,
-    val course: Int
+    val course: Double
 ): Parent
